android notif [nfc]: Fold NotificationHelper into NotificationUiManager
There's hardly anything left in this file, after #4842 made it
unnecessary to maintain our own data structure of conversations
because we can use the system's data structures of active
notifications.  Fold what little is left into the main file that
uses it (indeed the only file that uses its one nontrivial item.)

Also add some javadoc while we're at it.
